--- bash.exe.stackdump
@@ -1,9 +1,0 @@
-Stack trace:
-Frame        Function    Args
-000FFFFA310  00180062B0E (00180292068, 00180270E3E, 000FFFFA310, 000FFFF9210)
-000FFFFA310  0018004846A (00000000000, 00000000000, 00000000000, 00000000000)
-000FFFFA310  001800484A2 (00180292119, 000FFFFA1C8, 000FFFFA310, 00000000000)
-000FFFFA310  001800CEE9E (00000000000, 00000000000, 00000000000, 00000000000)
-000FFFFA310  001800CEFC5 (000FFFFA320, 00000000000, 00000000000, 00000000000)
-000FFFFA5E0  001800D0585 (000FFFFA320, 00000000000, 00000000000, 00000000000)
-End of stack trace
