--- my-app/.babelrc
@@ -1,6 +1,0 @@
-{
-  "presets": [
-      "@babel/preset-react"
-  ],
-  "plugins": ["@babel/plugin-syntax-jsx"]
-}
